requirements.txt
absl-py==2.0.0
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
appdirs==1.4.4
astunparse==1.6.3
asyncssh==2.14.2
atpublic==4.0
attrs==23.1.0
backoff==2.2.1
billiard==4.2.0
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.2
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.0.0
colorama==0.4.6
commonmark==0.9.1
configobj==5.0.8
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
dacite==1.6.0
dagshub==0.3.12
databricks-cli==0.18.0
dataclasses-json==0.6.3
dictdiffer==0.9.0
diskcache==5.6.3
distlib==0.3.8
distro==1.9.0
docker==6.1.3
dpath==2.1.6
dulwich==0.21.7
dvc==3.36.1
dvc-data==3.3.0
dvc-http==2.32.0
dvc-objects==3.0.0
dvc-render==1.0.0
dvc-studio-client==0.18.0
dvc-task==0.3.0
entrypoints==0.4
filelock==3.13.1
Flask==3.0.0
flatbuffers==23.5.26
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.47.0
frozenlist==1.4.1
fsspec==2023.12.2
funcy==2.0
fusepy==3.0.1
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
gql==3.4.1
grandalf==0.8
graphql-core==3.2.3
greenlet==3.0.3
grpcio==1.60.0
gto==1.6.1
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
httpcore==0.16.3
httpx==0.23.3
hydra-core==1.3.2
identify==2.5.33
idna==3.6
importlib-metadata==7.0.1
iterative-telemetry==0.0.8
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kaggle==1.5.16
keras==2.15.0
kiwisolver==1.4.5
kombu==5.3.4
libclang==16.0.6
loguru==0.7.2
Mako==1.3.0
Markdown==3.5.1
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
ml-dtypes==0.2.0
mlflow==2.9.2
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
nodeenv==1.8.0
numpy==1.26.2
oauthlib==3.2.2
omegaconf==2.3.0
opt-einsum==3.3.0
orjson==3.9.10
packaging==23.2
pandas==2.1.4
pathspec==0.12.1
pathvalidate==3.0.0
Pillow==10.1.0
platformdirs==3.11.0
polars==0.20.2
pre-commit==3.6.0
protobuf==4.25.1
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydot==1.4.2
pygit2==1.13.3
Pygments==2.17.2
pygtrie==2.5.0
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
querystring-parser==1.2.4
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3986==1.5.0
rich==13.1.0
rsa==4.9
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
scikit-learn==1.3.2
scipy==1.11.4
scmrepo==2.0.2
semver==3.0.2
shortuuid==1.0.11
shtab==1.6.5
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
sqltrie==0.11.0
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0.post1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.35.0
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.2.0
tomlkit==0.12.3
tqdm==4.66.1
treelib==1.6.4
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.18
vine==5.1.0
virtualenv==20.25.0
voluptuous==0.14.1
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
wrapt==1.14.1
yarl==1.9.4
zc.lockfile==3.0.post1
zipp==3.17.0
